+++ kwpbar/__init__.py
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
-# KwPBar for Python v0.1.1
+# KwPBar for Python v0.2.0
 # Copyright © 2015-2018, Chris Warrick.
 # All rights reserved.
 #
@@ -38,11 +38,12 @@ A progress bar for Python.
 :License: BSD (see /LICENSE).
 """
 
-import subprocess
+import shutil
+import subprocess  # used only on Python 2
 import sys
 
 __title__ = 'KwPBar for Python'
-__version__ = '0.1.1'
+__version__ = '0.2.0'
 __author__ = 'Chris Warrick'
 __license__ = '3-clause BSD'
 __docformat__ = 'restructuredtext en'
@@ -50,18 +51,22 @@ __docformat__ = 'restructuredtext en'
 __all__ = ('pbar', 'erase_pbar')
 
 
-def get_termlength():
+def get_termwidth(default=80):
     """Get the width of this terminal."""
     try:  # pragma: no cover
-        size = subprocess.check_output(['stty', 'size'])
-        return int(size.split()[1]) or 80
-    except (IndexError, subprocess.CalledProcessError):
-        return 80
+        return shutil.get_terminal_size((default, default)).columns
+    except AttributeError:
+        # Python 2
+        try:
+            size = subprocess.check_output(['stty', 'size'])
+            return int(size.split()[1]) or 80
+        except (IndexError, subprocess.CalledProcessError):
+            return 80
 
 
 def pbar(value, max):
     """Print the progressbar."""
-    fullwidth = get_termlength()
+    fullwidth = get_termwidth()
     pbarwidth = fullwidth - 9
     try:
         progress = float(value) / float(max)
@@ -84,7 +89,7 @@ def pbar(value, max):
 
 def erase_pbar():
     """Erase the progressbar."""
-    fullwidth = get_termlength()
+    fullwidth = get_termwidth()
     bar = ' ' * fullwidth
     sys.stderr.write("\r" + bar + "\r")
     sys.stderr.flush()