+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ Appendix C. Changelog
 :Author: Chris Warrick <chris@chriswarrick.com>
 :Copyright: © 2015-2018, Chris Warrick.
 :License: BSD (see /LICENSE or :doc:`Appendix B <LICENSE>`.)
-:Date: 2015-07-28
-:Version: 0.1.1
+:Date: 2018-09-03
+:Version: 0.2.0
 
 .. index:: CHANGELOG
 
@@ -19,11 +19,19 @@ More information can be found on GitHub in the `releases section
 Version History
 ===============
 
+0.2.0
+    * Replace subprocess call with shutil.get_terminal_size
+      (on compatible platforms)
+
 0.1.1
     * Add missing MANIFEST.in file
 
 0.1.0
     Initial release.
+
+0.2.0
+    * Replace subprocess call with shutil.get_terminal_size
+      (on compatible platforms)
 
 0.1.1
     * Add missing MANIFEST.in file